Requêtes mongodb

+++ cams.model/Core/PagedCollection.cs
@@ -3,7 +3,7 @@
 namespace cams.model.Core
 {
     /// <summary>
-    /// Collection with pagination information.
+    /// Defines the Collection with pagination information.
     /// </summary>
     /// <typeparam name="T">Object type.</typeparam>
     public class PagedCollection<T>
@@ -26,11 +26,11 @@ namespace cams.model.Core
         /// <summary>
         /// Gets or sets the number of items.
         /// </summary>
-        public int TotalNumberOfItems { get; set; }
+        public long TotalNumberOfItems { get; set; }
 
         /// <summary>
         /// Gets or sets the number of pages.
         /// </summary>
-        public int TotalNumberOfPages { get; set; }
+        public long TotalNumberOfPages { get; set; }
     }
 }
